requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
Flask
pyPDF2
python-docx
joblib
jsonify
scikit-learn